indexer/cert-indexer/requirements.txt
python-dotenv~=1.0
elasticsearch~=8.13
publicsuffixlist~=0.10
pika~=1.3